Create README
This file is an initial README to initially describe my code.  Currently it is the written out version of slide 3 of the project 2 pdf instructions.  The bounds of the workspace are defined using a lot of if-else statements.  I use slope to partially define the obstacle space of triangular sections.  The obstacle space is broken down into rectangles and right triangles in essence.  I did repeat something from the previous project, which is checking the previous move before looking through the entire closed list or open list.  Looking through those lists is expensive.  Checking if the current move is the opposite of the previous move is cheap if you record the move that resulted in the current node. The only other novel thing about my code is that I decided to use priority queues.