Update project2_functions.py
I added two functions to the end of this file.  The first function I wrote for project one to reverse the order of objects in a list. It is a while loop that goes through a list in reverse to add to a new list.  The second function is the trace back function used when the end goal is reached in project one.  It takes in the closed list, current node index, and current node parent.  It creates a list called path with those last two values.  Then a for loop loops through the closed list to find the node with the index that matches the parent node and updates the parent node with the found node's parent.  At the end it reverses the order of the list using the first function I mentioned.