Update project2_functions.py
I corrected a couple of the bounds in the function p2_coll.  I removed the commented out function p2_repeat.  I fixed the indexing of the priority queue in the traceback function (missing .queue in 3 spots).  I added a function called p2_obs.  This function is essentially the same as p2_coll, except the bounds are just the location of the obstacles.  This function will help me plot the obstacles and their bounds in different colors.