Add files via upload
This file was a test file.  The purpose of this file was to plot the obstacles and their buffers based on the moving object.  I had worried using a for loop to gather all these points would take a while, but this code runs in under a second.  I did not end up using the .csv files I created.  The way this file gathers the x and y coordinates is using a for loop that iterates through every point within the bounds of our workspace.  At each location it first checks if the point is located on an obstacle (p2_obs), and if not, it checks if the point is part of the buffer (p2_coll).  I plotted the points with a marker size of 1 to prevent overlap between the points.