Update project2_p1.py
When checking if the initial/goal nodes were within the bounds, I made use of the function exit if either are out of bounds.  Created the variable node_path (line 61) to utilize the trace back function I made previously. Removed the redundant 'check_lm == 0' from line 75.  Added a 'check_ob==1' to line 82.  Extracted all of the x and y points from the priority queues, and the node path.  I also added the code from project2_obs.py.  The points from the open loop and closed loop are green.  The obstacles are blue.  The buffer around the obstacles is red.  The final path to the node is purple and thicker than the rest.  Something is wrong with the node path, and I will have to fix that.  Otherwise, I need to animate this graph and make the initial/goal nodes user input to finish this code.