Update project2_p1.py
I found the issue with my code.  I obtained the x and y coordinates of the optimal path incorrectly with only one for loop to look through the parent index, but not matching it with x and y coordinates. I also a timer at the beginning and end of my code to see how long it would take to reach a point very far from the origin (530, 200).  My code took close to an hour to execute, although this is close to a worst case scenario.  Also, I added some code at the end of the file to save all of explored points and optimal path as .csv files.  I would like to be able to test my animation code quickly and loading in a .csv (proj2_q2.csv, proj2_q1.csv, and proj2_pa.csv) is a lot faster.  I will combine it later, but that combined file will be separate.