Add files via upload
This is my main source code for project 2.  It requires queue, matplotlib, and numpy to run.  However, it does not require any other python files.  It takes the user input for an initial position, goal position, and animation speed.  A good animation speed close to the start would be around 100, and one for a point very far from the start would 2000.  This is based off of how fast I believe an animation should run, and it is not an exact science.