Add files via upload
This file takes in the included .csv files to plot the path explored by the Dijkstra algorithm.  It is based on a start position of [10, 10], goal position of [530, 200], and an animation speed of 2000.  This was used for the .mp4 of this project as it cut down on downtime in the recording.